Fixed MeshlessAstar not getting the direct path
Added a workaround to Support in order to make it behave better with
IntCurvesFace_ShapeIntersector. The algorithm may take points at
different depths and that can cause problems. However, TopoDS_Face does
not behave well with SolidClassifier, so a long term solution will
involve somehow transforming it to a TopoDS_Shell. Will investigate
later on.