Fully add support for orthotropic materials
Beam generation has to account for the different behavior of orthotropic
materials while at an angle, due to its orthotropy, and maximum stresses
are also bound to be affected. So, change of coordinate system
transformations were added to account for that both in 3D and 2D.
Isotropic systems are, of course, much simpler and bound to be much
faster in this regard, as these rotations are unnecessary by definition.