Improved orthotropic maximum stress calculation
Adapted from Tsai-Wu criteria.

May have problems, as it is adapted from the calculation of principal
stress for normal and shear stresses separately. Still, beam sizing was
made assuming orthotropic materials, so it is not surprising it would
be a bit hard to adapt one to the other.

As sizing method looks for the greatest dimensions among the ones
calculated, it *should* be fine most of the time (that is, most of the
geometry). If it is to fail somewhere, it would probably be somewhere
where shear and normal stress are close (that is, close to the place
that the load is applied) and the geometry is not aligned to the axes.
If it proves to be a problem, summing the greatest of the normal and
shear dimensions obtained should probably be enough. This problem may
be less noticeable in 3D problems though, as the geometry is not as
sensitive to the loads in those cases.