Remove internal_loads variable from MeshNode
Using Node::results is fortunately sufficient, as there are equal
numbers of stress and load types per node, so get_internal_loads() and
get_stresses() can be used in alternation to get both. The current use
cases are not bound to use both at the same time, if it is necessary in
the future, it is possible to just create a new type of MeshNode for
that.