Change MeshElementFactory and GT9
Standardize constructor: takes positioning of values, ElementShape
instance and ProjectData pointer instead of being free to choose any
kind of constructor parameters.

Streamlines creation of instances of child class of MeshElement while
also making them relatively flexible (thanks to the input of ProjectData
pointer). Makes it is harder to use for prototyping a bit, but that can
be remedied with workaround methods if needed.