Modify method for calculating reaction moment
It's a more general and simpler method than the one used before, meaning
it also applies to indeterminate beams.

However, it obsoletes the iterative dimensional correction which was
used before. The method could still be used, but it probably would require
multiple finite element analysis instead of just correcting calculations
to account for distance change.

The code will remain for now, commented out.