Remove `results` variable from Node
It was barely used, and so used too much memory to be justifiable. It
was replaced by ad-hoc vectors which can be easily used and then
deleted.

Made necessary modifications. The few tests done indicate that sizing
step is still working correctly.

BeamSizing and related classes were also adapted, but are untested
(deprecated).