Refactor element and node creation
Element creation (MeshElementFactory) is now template based, further
automating its creation and making it less error prone. It does rely on
static variables, which may cause problems if they're missing. But does
problems are likely to be noticed at compile time due to it being based
on templates.

MeshNodeFactory was removed. The removal of Node::results eliminated the
need for different kinds of MeshNodes. Meshing still relies on a vector
of std::unique_ptr though, in order to preserve the pointers to the
nodes in case of changes to the vector (which does happen in
reordering).