Add MeshElementCommon interfaces
Partial implementations and specilizations of MeshElement, designed to
enhance code reuse while also making the process less error-prone and
easier to maintain, as it centralizes changes.

Should not increase overhead too much, as the implementation already
relied on polymorphism to work. It might, however, increase compilation
times, as it relies a lot on templating to work.