Implement definition of multiple mats and geoms
Full implementation will be done later, in order to ensure it works
correctly.

Mainly needs to modify behavior of creating and using elements for that,
and adapt DirectSolver/FiniteElement to take advantage of it (namely
caching D matrix and using it for multiple material topopt).

Adaptation of sizing algorithm will be postponed for now, as adapting
topology optimization and FEA is the priority for the following few
years.