Remove material and thickness from elements
Required data for calculations (namely, constitutive matrix `D` and
thickness `t`) are now requested by the functions that require them.

This provides some small benefits in terms of decreasing memory overhead
and decreasing CPU usage, as it removes two variables from each element
and does not require the `D` matrix to be recreated every time `get_k()`
or `get_DB()` is called.

These modifications also further the adaptation of the program towards
supporting multiple materials and geometries, as well as making it
easier to support topology optimization with two or more materials.