Make Meshing require a Geometry instance
The Geometry class is now what connects a shape and a material,
therefore, it is necessary to build stiffness matrices and such,
consequently being required for FiniteElement and subclasses.

This will simplify many things, although it makes a bit more difficult
to generate a finite element analysis from a given generated shape (as
is the case for StandardSizing).