Refactor FiniteElement and DirectSolver
Make the classes rely a lot less on ProjectData, also making it more
flexible and more reliable in terms of knowing what it's going to do
with the data it's supplied.

Change behavior on storage/generation of K to a more general
"step-based" approach, which should better suit future implementations
of iterative solvers.

Make the necessary changes.

Finite element analysis for a single geometry is still working. Possible
regressions, as well as behavior of FiniteElement, Meshing and
Visualization with multiple geometries and materials still need to be
investigated, but are likely to exist.