Fix filling Geometry::mesh being too slow
Other than that, it seems that meshing for multiple geometries and
merging nodes is working correctly. Current method is much faster, but
required limiting setting of internal boundary conditions for cases in
which there is only a single geometry, as there currently isn't a better
and simple alternative.

Will look into it later on if needed.