Make multiple materials actually work
FEA is working correctly.

MinimalVolume is working OK, although going for minimal volume without
void is not the best approach.

MinimalCompliance does not work at all, because it can't actually work
without void when restricting volume.

For both cases, going for mass instead of volume would give better
results. I might consider this approach in the future, as mass and
volume are equivalent when without void, but not when with it.