Add support for multiple materials in ViewHandler
Currently only works for two-material TO and such, as it's whats
supported.

Support for the nodal case was added, but might not look as expected.

Support for more than 2 materials (or 1 material + void) may not be
feasible using OneLab GUI as it does not support discrete per-element
coloring. An alternative may need to be developed for that.