Add script to generate Q4
Still needs some cleaning up to do and testing to do, but everything
looks ok. Unfornately, some calculations will need to be left to be done
numerically at runtime, as the symbolic equivalent proved too expensive.
That will result in some memory overhead to avoid CPU overhead, but it
should be ok.