Implement use of sorted nodes
There does not seem to be any regression, and misuse of
nodes/nodes_sorted is likely to cause some weirdnesses. I'll have to be
careful from now on when using them, but fortunately most of the further
additions necessary involve a lot of copy-pasting and reusing functions,
so maybe this way I'll be able to keep both of them and their uses in
mind.