Meshing: add generate_elements()
New function to unify and streamline the generation of mesh elements,
while also making it more efficient and just as extensible. This way, it
will also be more efficient to create elements and also to add other
meshers in the future if needed/desired.

Some functions may end up getting broken up, such as prepare_for_FEM(),
in order to make things easier to handle.

This is all in preparation for using boundary elements to generate the
force vector.

With this, it will also probably be possible to deprecate
StandardBeamMesher after this is ready, making things even easier.