Gmsh: rework boundary condition geometry
There was an attempt to apply the method of creating an special geometry
for boundary conditions in 2D, like it's being done in 3D.
Unfortunately, gmsh's square mesh algorithm does not respond well to
that, leading to it generating distorted meshes instead of regular ones.
Therefore, that has been reverted. A workaround may be needed for cubic
elements too.