BLAS: add BLIS optimizations
Initialize and finalize BLIS in main() if it is used. Makes it a bit
faster compared to its behavior without it, as bli_init_auto() can be a
performance bottleneck (according to perf).

Surprisingly, reference BLAS without -march-native seems to be the
fastest BLAS implementation for this use case. The main problem is that,
when using interpolation between different materials, lots of calls to
matrix inverses need to be done in parallel. As a result, there are tons
of calls to BLAS functions, initialization functions that are present in
them (like in BLIS) and thread locking (like in OpenBLAS) can
considerably hurt performance. Stock BLAS is pretty simple and
thread-local, so that is an advantage here.