Meshing: add conversion from frictionless to rigid
Another phase of implementation of frictionless contact: make initial
behavior equal to rigid and define mapping from rigid to non-rigid
versions by use of `node_positions` from SubProblem implementation.

`node_positions` is used internally to define the stifness matrix, and
nodes' internal `u_pos` is used on the global vectors to obtain and
display results, as well as in many parts of the calculations of
optimization functions. Altering `node_positions` to merge nodes for the
FEM problem solution without altering GlobalStiffnessMatrix or the
solvers does not seem to impact results and should be mathematically
sound. Currently, results seem to be unaltered compared to the rigid
version.