Add initial support for lambda-contact approaches
Being submit as a quite large commit as there are several interdependent
modifications and some architecture changes. Namely:

GlobalStiffnessMatrix is also responsible for generating the "expanded"
matrix. It still monopolizes matrix generation by use of
`generate_matrix_base()`, but the added `generate_expansion()` and
helper methods generate the correct matrix based on input.

Three types of matrices are currently defined: RIGID, LAMBDA_SLIDING
(only adds parallel lambdas) and LAMBDA_FULL. HESSIAN will implemented
as a modification of FULL which is copied from it.

FiniteElement is responsible both for linear and non-linear solving, but
subclasses only need to implement the "dumb" `solve()` method which is
used to solve linear problems. All non-linear procedures are based on
it.

SolverManager has barely changed, and is still mainly responsible for
setting up the problem and communicating results.

Meshing has been adapted to generate lambda elements (to be used for
matrix generating and tracking the global lambda vector) and also to
generate the final displacement vectors (using `expand_vector()` and
`apply_lambda()`).

Currently, the idea is to use LAMBDA_SLIDING as the matrix for getting
the reference displacement vector. However, that involves some problems:
(1) The global matrix is slightly bigger;
(2) Standard PCG does not seem to work well it for some reason. PETSc
says that the matrix becomes indefinite and with Eigen it diverges.
However, MUMPS and PETSc Cholesky can factorize it without any problem.

Therefore, using RIGID may actually be the best approach.

I also need to investigate whether it really is necessary to get the
initial `u` from solving the RIGID/LAMBDA_SLIDING problem or any
non-zero initial `u` would work for the Newton approach. This would
greatly simply things and require less memory.