Curvature: overhaul implementation
Use new energy-based implementation. It's still based on Lekhnitskii's
work, but is strain-based instead of stress-based, which makes it easier
to apply boundary conditions (if desired, although this is currently not
supported) and doesn't require solving a fourth-order PDE, as
compatibility conditions are already implicitly guaranteed.

This fixes a distortion that occurred in vertical flexural stress
(IIRC), and also makes the approach more mathematically sound.